Add GPU support to `phase_offset_correction`
The eigenvalue problem is still solved on CPU, but all other operations
are done on GPU. This means there is a transfer from/to GPU, and as such
a synchronization point. This should also be fine, as after this step,
we are probably mostly done with working on the large stack of complex
images.